+++ changelog.md
@@ -7,6 +7,7 @@ This version is not released yet and is under active development.
 ```
 
 - Add new `ANY_LINUX_COMPATIBILITY_LAYER` group.
+- Rename `click_extra.platform` to `click_extra.platforms`.
 
 ## {gh}`3.7.0 (2023-01-03) <compare/v3.6.0...v3.7.0>`